reset() is 1000's of times faster than memset

+++ cmd/relay/src/bench/bench.hpp
@@ -3,6 +3,7 @@
 
 #include <deque>
 #include <iostream>
+#include <cstddef>
 
 #include "clock.hpp"
 
@@ -35,7 +36,7 @@
 
 #define Do(times)  \
     Timer.reset(); \
-    for (int i = 0; i < times; i++)
+    for (size_t i = 0; i < times; i++)
 
 // Just for readability
 #define Skip()                                            \
